requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.23
langchain-openai>=0.0.5
python-dotenv==1.0.0 